requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argopt==0.8.2
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.7
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.12.0
git-fame==2.0.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
lxml==5.3.0
mammoth==1.9.0
Markdown @ file:///Users/ec2-user/ci_py311/markdown_1678377870854/work
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
plotly==5.24.1
posthog==3.11.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.2
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.1.28
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
st-chat-message==0.3.8
starlette==0.45.2
streamlit==1.41.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tkinterdnd2==0.4.2
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.17
unstructured-client==0.29.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
XlsxWriter==3.2.1
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
zstandard==0.23.0
